ppsutil/decoder.go:split NewPipelineManifestReader
- NewPipelineManifestReader now takes a byte slice instead of a path,
  and it's the caller's responsibility to read the path
