+++ src/internal/ppsutil/decoder.go
@@ -2,12 +2,7 @@ package ppsutil
 
 import (
 	"bytes"
-	"fmt"
 	"io"
-	"io/ioutil"
-	"net/http"
-	"net/url"
-	"os"
 
 	"gopkg.in/yaml.v3"
 
@@ -19,40 +14,13 @@ import (
 // PipelineManifestReader helps with unmarshalling pipeline configs from JSON.
 // It's used by 'create pipeline' and 'update pipeline'
 type PipelineManifestReader struct {
+	// Do first round of parsing (yaml -> holder) here (not serde), in case of
+	// multi-pipeline document
 	decoder *yaml.Decoder
 }
 
 // NewPipelineManifestReader creates a new manifest reader from a path.
-func NewPipelineManifestReader(path string) (result *PipelineManifestReader, retErr error) {
-	var pipelineBytes []byte
-	if path == "-" {
-		fmt.Print("Reading from stdin.\n")
-		var err error
-		pipelineBytes, err = ioutil.ReadAll(os.Stdin)
-		if err != nil {
-			return nil, err
-		}
-	} else if url, err := url.Parse(path); err == nil && url.Scheme != "" {
-		resp, err := http.Get(url.String())
-		if err != nil {
-			return nil, err
-		}
-		defer func() {
-			if err := resp.Body.Close(); err != nil && retErr == nil {
-				retErr = err
-			}
-		}()
-		pipelineBytes, err = ioutil.ReadAll(resp.Body)
-		if err != nil {
-			return nil, err
-		}
-	} else {
-		var err error
-		pipelineBytes, err = ioutil.ReadFile(path)
-		if err != nil {
-			return nil, err
-		}
-	}
+func NewPipelineManifestReader(pipelineBytes []byte) (result *PipelineManifestReader, retErr error) {
 	return &PipelineManifestReader{
 		decoder: yaml.NewDecoder(bytes.NewReader(pipelineBytes)),
 	}, nil
